Create AlphaBot_Coord
This version of AlphaBot was created 8-13-14 by Austin. It lets Robots 1 and 2 take turns sending distance data to Squadleader. Unfortunately, as of now it makes it very difficult to send commands to the robots (you have to click fast several times to get a command to go through). Delays should be experimented with to allow Squadleader to send commands easily.